backend/requirements.txt
Django
djangorestframework==3.14.0
psycopg2
python-clerk==0.1.0
webauthn==1.11.1
python-jose==3.3.0
cryptography==41.0.7
django-cors-headers==4.3.1
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1